changed all langauge from educator to user, implemented user table details + user table download. Need to work on User table upload and User add functionality.
