+++ webapp/src/screens/Admin.tsx
@@ -10,7 +10,7 @@ import Divider from "@mui/material/Divider";
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import ClassesTable from "./tabs/ClassesTable";
-import EducatorsTable from "./tabs/EducatorsTable";
+import UsersTable from "./tabs/UsersTable";
 
 const Admin = () => {
   const { screenState } = useContext(ScreenContext);
@@ -49,7 +49,7 @@ const Admin = () => {
                 <Tab label="Manage Classes" />
                 <Tab label="Manage Educators" />
               </Tabs>
-              { tabValue === 0 ? <ClassesTable /> : <EducatorsTable />}
+              { tabValue === 0 ? <ClassesTable /> : <UsersTable />}
             </Box>
           </Container>
         </Box>
