+++ webapp/src/utils/ExcelUtils.ts
@@ -1,6 +1,7 @@
 import {Exchange} from "../screens/Edit";
 import {Course} from "../components/onepirate/Home";
 import {CourseDoc} from "../screens/tabs/ClassesTable";
+import {UserDoc} from "../screens/tabs/UsersTable";
 
 export const spreadSheetData = [
   {
@@ -29,6 +30,21 @@ export const coursesSpreadSheetData = [
     content: [],
   }
 ]
+
+export const usersSpreadSheetData = [
+  {
+    sheet: "Users List",
+    columns: [
+      { label: "First Name", value: "firstName" },
+      { label: "Last Name", value: "lastName" },
+      { label: "Role IDs", value: "roleIds" },
+      { label: "Course IDs", value: "courseIds" },
+      { label: "Photo Link", value: "photoUrl" },
+    ],
+    content: [],
+  }
+]
+
 
 export interface ExcelJsonQuestions {
   numOfQuestions: number,
@@ -88,6 +104,38 @@ export const transformCoursesToJson: (stringArr: string[][]) => ExcelJsonCourses
   }
 }
 
+export interface ExcelJsonUsers {
+  numUsers: number,
+  users: Array<UserDoc>
+}
+
+// @ts-ignore
+export const transformUsersToJson: (stringArr: string[][]) => ExcelJsonUsers = (stringArr: string[][]) => {
+  if (stringArr.length === 0 || stringArr.length === 1) {
+    console.error("Invalid array given");
+    throw Error("Invalid array given");
+  }
+  const numUsers = stringArr.length -1;
+  const user: Array<UserDoc> = [];
+  stringArr.forEach((questionArray: string[], index) => {
+    if (index !== 0 && questionArray.length === 5) {
+      user.push({
+        firstName: questionArray[0],
+        lastName: questionArray[1],
+        roleIds: questionArray[2],
+        courseIds: questionArray[3],
+        photoUrl: questionArray[4],
+        id: "",
+        loginId: ""
+      })
+    }
+  });
+  return {
+    numCourses: numUsers,
+    courses: user
+  }
+}
+
 export const settings = {
   fileName: "QuestionsAnswersCS673", // Name of the resulting spreadsheet
   extraLength: 3, // A bigger number means that columns will be wider
@@ -98,6 +146,14 @@ export const settings = {
 
 export const courseExcelSettings = {
   fileName: "CoursesList", // Name of the resulting spreadsheet
+  extraLength: 3, // A bigger number means that columns will be wider
+  writeMode: "writeFile", // The available parameters are 'WriteFile' and 'write'. This setting is optional. Useful in such cases https://docs.sheetjs.com/docs/solutions/output#example-remote-file
+  writeOptions: {}, // Style options from https://docs.sheetjs.com/docs/api/write-options
+  RTL: false, // Display the columns from right-to-left (the default value is false)
+}
+
+export const userExcelSettings = {
+  fileName: "UsersList", // Name of the resulting spreadsheet
   extraLength: 3, // A bigger number means that columns will be wider
   writeMode: "writeFile", // The available parameters are 'WriteFile' and 'write'. This setting is optional. Useful in such cases https://docs.sheetjs.com/docs/solutions/output#example-remote-file
   writeOptions: {}, // Style options from https://docs.sheetjs.com/docs/api/write-options
