+++ webapp/src/utils/Urls.ts
@@ -20,12 +20,16 @@ export const getAllQnAUrl = () => ANSWERING_SVC_URL + "/exchanges/all";
 export const getAllCoursesUrl = () => ANSWERING_SVC_URL + "/courses";
 export const getAllCoursesForAdministrationUrl = () => ANSWERING_SVC_URL + "/courses/all";
 export const getAllDocumentsForCourseId = (courseId: string) => ANSWERING_SVC_URL + "/documents?courseId=" + courseId
+export const getAllUsersUrl = () => ANSWERING_SVC_URL + "/users";
 export const updateQuestionsUrl = () => ANSWERING_SVC_URL + "/exchanges";
 export const uploadDocumentsUrl = () => ANSWERING_SVC_URL + "/documents";
 export const addNewCourseUrl = () => ANSWERING_SVC_URL + "/courses";
+export const addNewUserUrl = () => ANSWERING_SVC_URL + "/courses";
 export const deleteCourseUrl = (courseId: string) => ANSWERING_SVC_URL + "/courses/" + courseId;
 export const deleteDocumentUrl = (documentId: string) => ANSWERING_SVC_URL + "/documents/" + documentId;
+export const deleteUserUrl = (educatorId: string) => ANSWERING_SVC_URL + "/users/" + educatorId;
 export const bulkUploadCoursesUrl = ()  =>  ANSWERING_SVC_URL + "/courses/all"
+export const bulkUploadUsersUrl = ()  =>  ANSWERING_SVC_URL + "/users"
 export const loginUrl = () => ANSWERING_SVC_URL + "/login";
 export const healthUrl = () => ANSWERING_SVC_URL + "/actuator/health";
 export const getOpenAIUrl = () => DOCUMENT_RETRIEVAL_SVC_URL + "/chat";
